remove BeanUtils and move mapToBean method to PojoUtils.mapToPojo
