+++ dubbo-cluster/src/main/java/org/apache/dubbo/rpc/cluster/router/tag/model/TagRuleParser.java
@@ -16,8 +16,8 @@
  */
 package org.apache.dubbo.rpc.cluster.router.tag.model;
 
-import org.apache.dubbo.common.utils.BeanUtils;
 import org.apache.dubbo.common.utils.CollectionUtils;
+import org.apache.dubbo.common.utils.PojoUtils;
 
 import org.yaml.snakeyaml.Yaml;
 import org.yaml.snakeyaml.constructor.SafeConstructor;
@@ -32,7 +32,7 @@ public class TagRuleParser {
     public static TagRouterRule parse(String rawRule) throws Exception {
         Yaml yaml = new Yaml(new SafeConstructor());
         Map<String, Object> map = yaml.load(rawRule);
-        TagRouterRule rule = BeanUtils.mapToBean(map, TagRouterRule.class);
+        TagRouterRule rule = PojoUtils.mapToPojo(map, TagRouterRule.class);
         rule.setRawRule(rawRule);
         if (CollectionUtils.isEmpty(rule.getTags())) {
             rule.setValid(false);
