+++ dubbo-common/src/main/java/org/apache/dubbo/common/utils/PojoUtils.java
@@ -23,6 +23,7 @@ import org.apache.dubbo.common.logger.LoggerFactory;
 import java.lang.reflect.Array;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
+import java.lang.reflect.GenericArrayType;
 import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
@@ -30,6 +31,8 @@ import java.lang.reflect.Modifier;
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Proxy;
 import java.lang.reflect.Type;
+import java.lang.reflect.TypeVariable;
+import java.lang.reflect.WildcardType;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
@@ -41,6 +44,7 @@ import java.util.IdentityHashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Properties;
 import java.util.TreeMap;
 import java.util.WeakHashMap;
@@ -49,6 +53,8 @@ import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.ConcurrentSkipListMap;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
+
+import static org.apache.dubbo.common.utils.ClassUtils.isAssignableFrom;
 
 /**
  * PojoUtils. Travel object deeply, and convert complex type to simple type.
@@ -68,6 +74,8 @@ public class PojoUtils {
     private static final ConcurrentMap<String, Method> NAME_METHODS_CACHE = new ConcurrentHashMap<String, Method>();
     private static final ConcurrentMap<Class<?>, ConcurrentMap<String, Field>> CLASS_FIELD_CACHE = new ConcurrentHashMap<Class<?>, ConcurrentMap<String, Field>>();
     private static final boolean GENERIC_WITH_CLZ = Boolean.parseBoolean(ConfigUtils.getProperty(CommonConstants.GENERIC_WITH_CLZ_KEY, "true"));
+    private static final List<Class<?>> CLASS_CAN_BE_STRING = Arrays.asList(Byte.class, Short.class, Integer.class,
+            Long.class, Float.class, Double.class, Boolean.class, Character.class);
 
     public static Object[] generalize(Object[] objs) {
         Object[] dests = new Object[objs.length];
@@ -674,6 +682,95 @@ public class PojoUtils {
     public static <T> void updatePropertyIfAbsent(Supplier<T> getterMethod, Consumer<T> setterMethod, T newValue) {
         if (newValue != null && getterMethod.get() == null) {
             setterMethod.accept(newValue);
+        }
+    }
+
+    /**
+     * convert map to a specific class instance
+     *
+     * @param map map wait for convert
+     * @param cls the specified class
+     * @param <T> the type of {@code cls}
+     * @return class instance declare in param {@code cls}
+     * @throws ReflectiveOperationException if the instance creation is failed
+     * @since 2.7.10
+     */
+    public static <T> T mapToPojo(Map<String, Object> map, Class<T> cls) throws ReflectiveOperationException {
+        T instance = cls.getDeclaredConstructor().newInstance();
+        Map<String, Field> beanPropertyFields = ReflectUtils.getBeanPropertyFields(cls);
+        for (Map.Entry<String, Field> entry : beanPropertyFields.entrySet()) {
+            String name = entry.getKey();
+            Field field = entry.getValue();
+            Object mapObject = map.get(name);
+            if (mapObject == null) {
+                continue;
+            }
+
+            Type type = field.getGenericType();
+            Object fieldObject = getFieldObject(mapObject, type);
+            field.set(instance, fieldObject);
+        }
+
+        return instance;
+    }
+
+    private static Object getFieldObject(Object mapObject, Type fieldType) throws ReflectiveOperationException {
+        if (fieldType instanceof Class<?>) {
+            return convertClassType(mapObject, (Class<?>) fieldType);
+        } else if (fieldType instanceof ParameterizedType) {
+            return convertParameterizedType(mapObject, (ParameterizedType) fieldType);
+        } else if (fieldType instanceof GenericArrayType || fieldType instanceof TypeVariable<?> || fieldType instanceof WildcardType) {
+            // ignore these type currently
+            return null;
+        } else {
+            throw new IllegalArgumentException("Unrecognized Type: " + fieldType.toString());
+        }
+    }
+
+    @SuppressWarnings("unchecked")
+    private static Object convertClassType(Object mapObject, Class<?> type) throws ReflectiveOperationException {
+        if (type.isPrimitive() || isAssignableFrom(type, mapObject.getClass())) {
+            return mapObject;
+        } else if (Objects.equals(type, String.class) && CLASS_CAN_BE_STRING.contains(mapObject.getClass())) {
+            // auto convert specified type to string
+            return mapObject.toString();
+        } else if (mapObject instanceof Map) {
+            return mapToPojo((Map<String, Object>) mapObject, type);
+        } else {
+            // type didn't match and mapObject is not another Map struct.
+            // we just ignore this situation.
+            return null;
+        }
+    }
+
+    @SuppressWarnings("unchecked")
+    private static Object convertParameterizedType(Object mapObject, ParameterizedType type) throws ReflectiveOperationException {
+        Type rawType = type.getRawType();
+        if (!isAssignableFrom((Class<?>) rawType, mapObject.getClass())) {
+            return null;
+        }
+
+        Type[] actualTypeArguments = type.getActualTypeArguments();
+        if (isAssignableFrom(Map.class, (Class<?>) rawType)) {
+            Map<Object, Object> map = (Map<Object, Object>) mapObject.getClass().getDeclaredConstructor().newInstance();
+            for (Map.Entry<Object, Object> entry : ((Map<Object, Object>) mapObject).entrySet()) {
+                Object key = getFieldObject(entry.getKey(), actualTypeArguments[0]);
+                Object value = getFieldObject(entry.getValue(), actualTypeArguments[1]);
+                map.put(key, value);
+            }
+
+            return map;
+        } else if (isAssignableFrom(Collection.class, (Class<?>) rawType)) {
+            Collection<Object> collection = (Collection<Object>) mapObject.getClass().getDeclaredConstructor().newInstance();
+            for (Object m : (Iterable<?>) mapObject) {
+                Object ele = getFieldObject(m, actualTypeArguments[0]);
+                collection.add(ele);
+            }
+
+            return collection;
+        } else {
+            // ignore other type currently
+            return null;
         }
     }
